requirements.txt
attrs==23.2.0
Automat==22.10.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
cssselect==1.2.0
filelock==3.13.3
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
lxml==5.2.1
mysql-connector-python==8.3.0
packaging==24.0
parsel==1.9.0
Protego==0.3.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==24.1.0
queuelib==1.6.2
requests==2.31.0
requests-file==2.0.0
Scrapy==2.11.1
service-identity==24.1.0
setuptools==69.2.0
six==1.16.0
soupsieve==2.5
tldextract==5.1.2
Twisted==24.3.0
typing_extensions==4.10.0
urllib3==2.2.1
w3lib==2.1.2
zope.interface==6.2
